Makefile: Silence warnings about array sensitivity
Icarus verilog complains if you are sensitive to every element in an
array:

rtl/mii_elastic_buffer.v:78: warning: @* is sensitive to all 5 words in array 'data'.

This makes sense if you intend to synthesize this array to a block RAM,
but not really if it's supposed to be registers. Silence this warning.

Signed-off-by: Sean Anderson <seanga2@gmail.com>